+++ NEWS.md
@@ -1,5 +1,7 @@
 # waldo (development version)
 
+* `compare()` now goes into more details if you're comparing an S3 object with a base object of the same type (#218).
+
 # waldo 0.6.1
 
 * Only use special bit64 comparison if package is installed.